Majorly optimize memory and CPU usage in file loading
Previously, the file loading code made copies of the file in RAM,
resulting in quadratic CPU and memory usage. This simple fix makes file
loading linear, as expected.

TODO: In the future, this could be optimized for even less copying, by not
loading the file into an intermediate Vec<u8>, and feeding the file contents
directly into the TreeBuilder.